// frontend/src/types/questionBank.ts
export interface BankQuestionOption {
    key: 'A' | 'B' | 'C' | 'D';
    text_en: string;
    text_bn: string;
    imageUrl?: string;
}